+++ CHANGES.md
@@ -32,6 +32,7 @@
   * Super NES: Satellaview BS-X ROM headers are now decoded properly.
   * Wii: Unencrypted images from RVT-H Reader systems are now supported.
   * NES: Improved publisher lookup for FDS titles.
+  * Wii U: Display game publishers.
 
 * New compressed texture formats:
   * Ericsson ETC1 and ETC2